# openapi/openapi.yaml
openapi: 3.1.0
info: { title: 趣然 AI 商品经营与利润决策助手 API, version: 1.0.0 }
paths:
  /api/session:
    get:
      operationId: getSession
      responses:
        '200': { description: 当前会话, content: { application/json: { schema: { $ref: '#/components/schemas/SessionResponse' } } } }
        '401': { $ref: '#/components/responses/Error' }
  /api/batches:
    get:
      operationId: listBatches
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: limit, in: query, schema: { type: integer, enum: [20, 50, 100] } }
      responses:
        '200': { description: 批次分页列表, content: { application/json: { schema: { $ref: '#/components/schemas/BatchListResponse' } } } }
    post:
      operationId: createBatch
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [business_unit, period_start, period_end, business_cutoff_date, xlsx_file]
              properties:
                business_unit: { type: string, const: 玩具事业部 }
                period_start: { type: string, format: date }
                period_end: { type: string, format: date }
                business_cutoff_date: { type: string, format: date }
                xlsx_file: { type: string, format: binary }
      responses:
        '201': { description: 新批次, content: { application/json: { schema: { $ref: '#/components/schemas/BatchSummary' } } } }
        '200': { description: 幂等命中, content: { application/json: { schema: { $ref: '#/components/schemas/BatchSummary' } } } }
        '422': { $ref: '#/components/responses/Error' }
  /api/batches/{batchId}:
    get:
      operationId: getBatch
      parameters:
        - { name: batchId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: 批次详情, content: { application/json: { schema: { $ref: '#/components/schemas/BatchDetail' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /api/actions:
    get:
      operationId: listActions
      parameters:
        - { name: batch_id, in: query, schema: { type: string, format: uuid } }
        - { name: search, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string } }
        - { name: store, in: query, schema: { type: string } }
        - { name: operator, in: query, schema: { type: string } }
        - { name: review_status, in: query, schema: { type: string } }
        - { name: business_state, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: limit, in: query, schema: { type: integer, enum: [20, 50, 100] } }
      responses:
        '200': { description: 行动清单, content: { application/json: { schema: { $ref: '#/components/schemas/ActionListResponse' } } } }
        '401': { $ref: '#/components/responses/Error' }
  /api/workbench:
    get:
      operationId: getWorkbench
      responses:
        '200': { description: 当前角色工作台, content: { application/json: { schema: { $ref: '#/components/schemas/Workbench' } } } }
        '401': { $ref: '#/components/responses/Error' }
  /api/suggestions/{linkId}:
    get:
      operationId: getSuggestion
      parameters: [{ name: linkId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: 建议详情, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /api/suggestions/{linkId}/review:
    post:
      operationId: reviewSuggestion
      parameters: [{ name: linkId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReviewInput' } } }
      responses:
        '200': { description: 审核后的建议详情, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /api/suggestions/{linkId}/override:
    post:
      operationId: overrideSuggestion
      parameters: [{ name: linkId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OverrideInput' } } } }
      responses:
        '200': { description: 改判后的建议详情, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /api/suggestions/{linkId}/terminate:
    post:
      operationId: terminateSuggestion
      parameters: [{ name: linkId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TerminateInput' } } } }
      responses:
        '200': { description: 终止后的建议详情, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /api/actions/{taskId}/execute:
    post:
      operationId: executeAction
      parameters: [{ name: taskId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ExecuteInput' } } } }
      responses:
        '200': { description: 执行后的建议, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /api/actions/{taskId}/result:
    post:
      operationId: recordActionResult
      parameters: [{ name: taskId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ResultInput' } } } }
      responses:
        '200': { description: 记录结果后的建议, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /api/actions/{taskId}/clearance-completion:
    post:
      operationId: submitClearanceCompletion
      parameters: [{ name: taskId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClearanceSubmitInput' } } } }
      responses:
        '200': { description: 提交后的清仓完成详情, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /api/actions/{taskId}/confirm:
    post:
      operationId: confirmClearanceCompletion
      parameters: [{ name: taskId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClearanceDecisionInput' } } } }
      responses:
        '200': { description: 已确认的清仓完成详情, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /api/actions/{taskId}/return:
    post:
      operationId: returnClearanceCompletion
      parameters: [{ name: taskId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ClearanceDecisionInput' } } } }
      responses:
        '200': { description: 已退回的清仓完成详情, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /api/actions/{taskId}/oa-notifications:
    post:
      operationId: sendOANotification
      parameters: [{ name: taskId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OANotificationInput' } } } }
      responses:
        '200': { description: OA 发送结果及建议详情, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
  /api/actions/{taskId}/oa-notifications/{notificationId}/retry:
    post:
      operationId: retryOANotification
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: notificationId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OARetryInput' } } } }
      responses:
        '200': { description: OA 补发结果及建议详情, content: { application/json: { schema: { $ref: '#/components/schemas/SuggestionDetail' } } } }
        '409': { $ref: '#/components/responses/Error' }
components:
  responses:
    Error: { description: 业务错误, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  schemas:
    BusinessRole: { type: string, enum: [operations, supervisor] }
...
    SessionResponse:
      type: object
      required: [authenticated, user]
      properties: { authenticated: { type: boolean, const: true }, user: { $ref: '#/components/schemas/SessionUser' } }
...
    BatchSummary:
      type: object
      required: [id, code, business_unit, period_start, period_end, business_cutoff_date, source_file_name, status, valid_count, rejected_count, degraded_count, warning_count, rule_version, failure_code, created_by, created_at, completed_at]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        business_unit: { type: string }
        period_start: { type: string, format: date }
        period_end: { type: string, format: date }
        business_cutoff_date: { type: string, format: date }
        source_file_name: { type: string }
        status: { $ref: '#/components/schemas/BatchStatus' }
        valid_count: { type: [integer, 'null'] }
        rejected_count: { type: [integer, 'null'] }
        degraded_count: { type: [integer, 'null'] }
        warning_count: { type: [integer, 'null'] }
        rule_version: { type: [string, 'null'] }
        failure_code: { type: [string, 'null'] }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        idempotent: { type: boolean }
...
    BatchDetail:
      allOf:
        - { $ref: '#/components/schemas/BatchSummary' }
        - type: object
          required: [issues, snapshots]
          properties:
            issues: { type: array, items: { $ref: '#/components/schemas/ImportIssue' } }
            snapshots: { type: array, items: { $ref: '#/components/schemas/Snapshot' } }
    BatchListResponse:
      type: object
      required: [items, page, limit, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/BatchSummary' } }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
...
    ActionListResponse:
      type: object
      required: [items, page, limit, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ActionItem' } }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
    Workbench:
      type: object
      required: [role, latest_batch_id, latest_batch_code, batch_completed_at, pending_review_count, pending_execution_count, clearance_confirm_count, exception_count, items]
      properties:
        role: { $ref: '#/components/schemas/BusinessRole' }
        latest_batch_id: { type: string }
        latest_batch_code: { type: string }
        batch_completed_at: { type: string, format: date-time }
        pending_review_count: { type: integer }
        pending_execution_count: { type: integer }
        clearance_confirm_count: { type: integer }
        exception_count: { type: integer }
        items: { type: array, items: { $ref: '#/components/schemas/ActionItem' } }
...
    SuggestionDetail:
      allOf:
        - { $ref: '#/components/schemas/ActionItem' }
        - type: object
          required: [events, ai_status, ai_content, clearance_completion, notifications]
          properties:
            events: { type: array, items: { $ref: '#/components/schemas/BusinessEvent' } }
            ai_status: { type: string }
            ai_content: { type: object, additionalProperties: true }
            clearance_completion: { anyOf: [{ $ref: '#/components/schemas/ClearanceCompletion' }, { type: 'null' }] }
            notifications: { type: array, items: { $ref: '#/components/schemas/OANotification' } }
...
    ReviewInput:
      type: object
      required: [decision, note, review_version, idempotency_key]
      properties:
        decision: { type: string, enum: [approved, rejected] }
        note: { type: string }
        review_version: { type: integer, minimum: 1 }
        idempotency_key: { type: string, minLength: 1 }
    ExecuteInput:
      type: object
      required: [track, version, note, idempotency_key]
      properties:
        track: { type: string, enum: [business, inventory] }
        version: { type: integer, minimum: 1 }
        note: { type: string }
        idempotency_key: { type: string, minLength: 1 }
    ResultInput:
      type: object
      required: [period_start, period_end, sales_value, profit_value, inventory_value, sales_unavailable, profit_unavailable, inventory_unavailable, note, version, idempotency_key]
      properties:
        period_start: { type: string, format: date }
        period_end: { type: string, format: date }
        sales_value: { type: [number, 'null'] }
        profit_value: { type: [number, 'null'] }
        inventory_value: { type: [number, 'null'] }
        sales_unavailable: { type: boolean }
        profit_unavailable: { type: boolean }
        inventory_unavailable: { type: boolean }
        note: { type: string, minLength: 1 }
        version: { type: integer, minimum: 1 }
        idempotency_key: { type: string, minLength: 1 }
    OverrideInput:
      type: object
      required: [business_action, inventory_action, reason, version, idempotency_key]
      properties:
        business_action: { type: string, enum: [clearance, stop_loss, observe, invest, maintain] }
        inventory_action: { type: [string, 'null'], enum: [restock, no_restock, prohibit_restock, null] }
        reason: { type: string, minLength: 1 }
        version: { type: integer, minimum: 1 }
        idempotency_key: { type: string, minLength: 1 }
    TerminateInput:
      type: object
      required: [reason, version, idempotency_key]
      properties:
        reason: { type: string, minLength: 1 }
        version: { type: integer, minimum: 1 }
        idempotency_key: { type: string, minLength: 1 }
...
    ClearanceDecisionInput:
      type: object
      required: [reason, version, idempotency_key]
      properties:
        reason: { type: string }
        version: { type: integer, minimum: 1 }
        idempotency_key: { type: string, minLength: 1 }
...